Rewrite handling of legacy formatted text and add tests
