+++ Sources/Core/Sources/Chat/LegacyFormattedTextStyle.swift
@@ -1,0 +1,10 @@
+extension LegacyFormattedText {
+  /// A legacy text style formatting code.
+  public enum Style: Character {
+    case bold = "l"
+    case strikethrough = "m"
+    case underline = "n"
+    case italic = "o"
+    case reset = "r"
+  }
+}
