+++ Sources/Core/Sources/Chat/LegacyFormattedTextToken.swift
@@ -1,0 +1,8 @@
+extension LegacyFormattedText {
+  /// A formatted text token.
+  public struct Token: Equatable {
+    let string: String
+    let color: Color?
+    let style: Style?
+  }
+}
